Add dummy tun, helper functions
