+++ src/yggdrasil/tun_dummy.go
@@ -1,8 +1,6 @@
-// +build !linux,!darwin,!windows,!openbsd,!freebsd,!netbsd,!mobile
+// +build mobile
 
 package yggdrasil
-
-import water "github.com/yggdrasil-network/water"
 
 // This is to catch unsupported platforms
 // If your platform supports tun devices, you could try configuring it manually
@@ -10,17 +8,6 @@ import water "github.com/yggdrasil-network/water"
 // Creates the TUN/TAP adapter, if supported by the Water library. Note that
 // no guarantees are made at this point on an unsupported platform.
 func (tun *tunAdapter) setup(ifname string, iftapmode bool, addr string, mtu int) error {
-	var config water.Config
-	if iftapmode {
-		config = water.Config{DeviceType: water.TAP}
-	} else {
-		config = water.Config{DeviceType: water.TUN}
-	}
-	iface, err := water.New(config)
-	if err != nil {
-		panic(err)
-	}
-	tun.iface = iface
 	tun.mtu = getSupportedMTU(mtu)
 	return tun.setupAddress(addr)
 }
@@ -28,6 +15,5 @@ func (tun *tunAdapter) setup(ifname string, iftapmode bool, addr string, mtu int
 // We don't know how to set the IPv6 address on an unknown platform, therefore
 // write about it to stdout and don't try to do anything further.
 func (tun *tunAdapter) setupAddress(addr string) error {
-	tun.core.log.Println("Platform not supported, you must set the address of", tun.iface.Name(), "to", addr)
 	return nil
 }
